@@ -9,14 +9,19 @@
   },
   "author": "Tyler McGinnis",
   "dependencies": {
-    "events": "^1.0.2",
-    "firebase": "^2.2.0",
-    "flux": "^2.0.1",
-    "react": "^0.12.2",
-    "react-router": "^0.12.0"
+    "events": "*",
+    "firebase": "^2.2.3",
+    "flux": "*",
+    "react": "^0.13.1",
+    "react-router": "^0.13.1"
   },
   "devDependencies": {
-    "jsx-loader": "^0.12.2",
-    "webpack": "^1.5.3"
+    "babel": "*",
+    "babel-core": "*",
+    "babel-loader": "*",
+    "jsx-loader": "*",
+    "react-hot-loader": "*",
+    "webpack": "*",
+    "webpack-dev-server": "*"
   }
 }
