@@ -1,9 +1,9 @@
 var React = require("react");
 
-var Home = React.createClass({
-  render: function(){
-    return <p> This is the Home Page. This route is not protected. </p>
+class Home extends React.Component{
+  render(){
+    return <p> This is the WOOO Page. This route is not protected. </p>
   }
-});
+};
 
 module.exports = Home;
